+++ IBAnimatable/ActivityIndicatorAnimationBallClipRotate.swift
@@ -50,7 +50,7 @@ private extension ActivityIndicatorAnimationBallClipRotate {
   var animation: CAAnimationGroup {
     let animation = CAAnimationGroup()
     animation.animations = [scaleAnimation, rotateAnimation]
-    animation.timingFunction = CAMediaTimingFunction(name: kCAMediaTimingFunctionLinear)
+    animation.timingFunctionType = .linear
     animation.duration = duration
     animation.repeatCount = .infinity
     animation.isRemovedOnCompletion = false