+++ IBAnimatable/AnimatableTableView.swift
@@ -152,6 +152,12 @@ open var startPoint: GradientStartPoint = .top
   @IBInspectable open var damping: CGFloat = CGFloat.nan
   @IBInspectable open var velocity: CGFloat = CGFloat.nan
   @IBInspectable open var force: CGFloat = CGFloat.nan
+  @IBInspectable var _timingFunction: String = "" {
+    didSet {
+      timingFunction = TimingFunctionType(string: _timingFunction)
+    }
+  }
+  open var timingFunction: TimingFunctionType = .none
 
   // MARK: - Lifecycle
   open override func prepareForInterfaceBuilder() {