+++ IBAnimatable/AnimatableTextView.swift
@@ -113,6 +113,12 @@ open class AnimatableTextView: UITextView, CornerDesignable, FillDesignable, Bor
   @IBInspectable open var damping: CGFloat = CGFloat.nan
   @IBInspectable open var velocity: CGFloat = CGFloat.nan
   @IBInspectable open var force: CGFloat = CGFloat.nan
+  @IBInspectable var _timingFunction: String = "" {
+    didSet {
+      timingFunction = TimingFunctionType(string: _timingFunction)
+    }
+  }
+  open var timingFunction: TimingFunctionType = .none
 
   // MARK: Override properties
   override open var font: UIFont! {