+++ IBAnimatable/AnimationChainable.swift
@@ -14,6 +14,7 @@ struct AnimationConfiguration {
   let duration: TimeInterval
   let delay: TimeInterval
   let force: CGFloat
+  let timingFunction: TimingFunctionType
 }
 
 public class AnimationPromise<T: UIView> where T: Animatable {
@@ -45,12 +46,14 @@ public class AnimationPromise<T: UIView> where T: Animatable {
                     duration: TimeInterval? = nil,
                     damping: CGFloat? = nil,
                     velocity: CGFloat? = nil,
-                    force: CGFloat? = nil) -> AnimationPromise {
+                    force: CGFloat? = nil,
+                    timingFunction: TimingFunctionType? = nil) -> AnimationPromise {
     let configuration = AnimationConfiguration(damping: damping ?? view.damping,
                                                velocity: velocity ?? view.velocity,
                                                duration: duration ?? view.duration,
                                                delay: delayForNextAnimation,
-                                               force: force ?? view.force)
+                                               force: force ?? view.force,
+                                               timingFunction: timingFunction ?? view.timingFunction)
     let animTuple = AnimationTuple(type: animation, configuration: configuration)
     animationList.append(animTuple)
     if animationList.count == 1 { // If it's the only animation, launch it immediately