+++ IBAnimatableTests/CornerDesignableTests.swift
@@ -2,80 +2,27 @@
 //  CornerDesignableTests.swift
 //  IBAnimatable
 //
-//  Created by Steven on 4/24/17.
+//  Created by Steven on 5/12/17.
 //  Copyright © 2017 IBAnimatable. All rights reserved.
 //
 
 import XCTest
 @testable import IBAnimatable
 
-final class CornerDesignableTests: XCTestCase {
+// MARK: - CornerDesignableTests Protocol
 
-  // MARK: - testCornerRadius
+protocol CornerDesignableTests: class {
 
-  func testCornerRadius() {
-    helper_testCornerRadius_UIView(element: AnimatableView())
-    helper_testCornerRadius_UIView(element: AnimatableCheckBox())
-    helper_testCornerRadius_UIView(element: AnimatableImageView())
-    helper_testCornerRadius_UIView(element: AnimatableLabel())
-    helper_testCornerRadius_UIView(element: AnimatableStackView())
-    helper_testCornerRadius_UIView(element: AnimatableTableViewCell())
-    helper_testCornerRadius_UIView(element: AnimatableStackView())
-    helper_testCornerRadius_UIView(element: AnimatableView())
-    helper_testCornerRadius_UICollectionViewCell(element: AnimatableCollectionViewCell())
-  }
+  func testCornerRadius()
+  func test_cornerSides()
 
-  private func helper_testCornerRadius_UIView(element: MockUIViewCornerDesignable) {
-    element.cornerRadius = 3.0
-    element.cornerSides = .allSides
-    XCTAssertEqual(element.cornerRadius, element.layer.cornerRadius)
-    element.cornerSides = [.bottomLeft, .bottomRight, .topLeft]
-    let mask = element.layer.mask as? CAShapeLayer
-    XCTAssertEqual(mask?.frame, CGRect(origin: .zero, size: element.bounds.size))
-    XCTAssertEqual(mask?.name, "cornerSideLayer")
-    let mockPath = cgPath(for: element, withCorners: [.bottomLeft, .bottomRight, .topLeft])
-    XCTAssertEqual(mask?.path, mockPath)
-  }
+}
 
-  private func helper_testCornerRadius_UICollectionViewCell(element: MockUICollectionViewCellCornerDesignable) {
-    element.cornerRadius = -1
-    XCTAssertFalse(element.contentView.layer.masksToBounds)
-    element.cornerRadius = 3.0
-    XCTAssertNil(element.layer.mask)
-    XCTAssertEqual(element.layer.cornerRadius, 0.0)
-    element.cornerSides = .allSides
-    XCTAssertEqual(element.contentView.layer.cornerRadius, element.cornerRadius)
-    element.cornerSides = [.bottomLeft, .bottomRight, .topLeft]
-    XCTAssertEqual(element.contentView.layer.cornerRadius, 0.0)
-    let mask = element.contentView.layer.mask as? CAShapeLayer
-    XCTAssertEqual(mask?.frame, CGRect(origin: .zero, size: element.bounds.size))
-    XCTAssertEqual(mask?.name, "cornerSideLayer")
-    let mockPath = cgPath(for: element, withCorners: [.bottomLeft, .bottomRight, .topLeft])
-    XCTAssertEqual(mask?.path, mockPath)
-    XCTAssertTrue(element.contentView.layer.masksToBounds)
-  }
+// MARK: - Universal Tests
 
-  private func cgPath(for element: MockUIViewCornerDesignable, withCorners corners: UIRectCorner) -> CGPath {
-    let cornerRadii = CGSize(width: element.cornerRadius, height: element.cornerRadius)
-    return UIBezierPath(roundedRect: element.bounds, byRoundingCorners: corners, cornerRadii: cornerRadii).cgPath
-  }
+extension CornerDesignableTests {
 
-  // MARK: - test_cornerSides
-
-  func test_cornerSides() {
-    helper_test_cornerSides(element: AnimatableButton())
-    helper_test_cornerSides(element: AnimatableCheckBox())
-    helper_test_cornerSides(element: AnimatableCollectionViewCell())
-    helper_test_cornerSides(element: AnimatableImageView())
-    helper_test_cornerSides(element: AnimatableLabel())
-    helper_test_cornerSides(element: AnimatableStackView())
-    helper_test_cornerSides(element: AnimatableTableViewCell())
-    helper_test_cornerSides(element: AnimatableTextField())
-    helper_test_cornerSides(element: AnimatableTextView())
-    helper_test_cornerSides(element: AnimatableView())
-  }
-
-  private func helper_test_cornerSides(element: MockUIViewCornerDesignable) {
+  func _test_cornerSides(_ element: StringCornerDesignable) {
     element._cornerSides = "topLeft"
     XCTAssertEqual(element.cornerSides, .topLeft)
     element._cornerSides = "topRight"
@@ -92,31 +39,48 @@ final class CornerDesignableTests: XCTestCase {
 
 }
 
-// MARK: - MockUICornerDesignable
-fileprivate protocol MockUIViewCornerDesignable: class, CornerDesignable {
+// MARK: - UIView Tests
 
-  var _cornerSides: String? { get set }
-  var layer: CALayer { get }
-  var bounds: CGRect { get set }
+extension CornerDesignableTests {
+
+  func _testCornerRadius<E: UIView>(_ element: E) where E: CornerDesignable {
+    element.cornerRadius = 3.0
+    element.cornerSides = .allSides
+    XCTAssertEqual(element.cornerRadius, element.layer.cornerRadius)
+    element.cornerSides = [.bottomLeft, .bottomRight, .topLeft]
+    let mask = element.layer.mask as? CAShapeLayer
+    XCTAssertEqual(mask?.frame, CGRect(origin: .zero, size: element.bounds.size))
+    XCTAssertEqual(mask?.name, "cornerSideLayer")
+    let cornerRadii = CGSize(width: element.cornerRadius, height: element.cornerRadius)
+    let corners: UIRectCorner = [.bottomLeft, .bottomRight, .topLeft]
+    let mockPath = UIBezierPath(roundedRect: element.bounds, byRoundingCorners: corners, cornerRadii: cornerRadii).cgPath
+    XCTAssertEqual(mask?.path, mockPath)
+  }
 
 }
 
-// MARK: - MockUICollectionViewCellCornerDesignable
-fileprivate protocol MockUICollectionViewCellCornerDesignable: class, MockUIViewCornerDesignable {
+// MARK: - UICollectionViewCell Tests
 
-  var contentView: UIView { get }
+extension CornerDesignableTests {
+
+  func _testCornerRadius<E: UICollectionViewCell>(_ element: E) where E: CornerDesignable {
+    element.cornerRadius = -1
+    XCTAssertFalse(element.contentView.layer.masksToBounds)
+    element.cornerRadius = 3.0
+    XCTAssertNil(element.layer.mask)
+    XCTAssertEqual(element.layer.cornerRadius, 0.0)
+    element.cornerSides = .allSides
+    XCTAssertEqual(element.contentView.layer.cornerRadius, element.cornerRadius)
+    element.cornerSides = [.bottomLeft, .bottomRight, .topLeft]
+    XCTAssertEqual(element.contentView.layer.cornerRadius, 0.0)
+    let mask = element.contentView.layer.mask as? CAShapeLayer
+    XCTAssertEqual(mask?.frame, CGRect(origin: .zero, size: element.bounds.size))
+    XCTAssertEqual(mask?.name, "cornerSideLayer")
+    let cornerRadii = CGSize(width: element.cornerRadius, height: element.cornerRadius)
+    let corners: UIRectCorner = [.bottomLeft, .bottomRight, .topLeft]
+    let mockPath = UIBezierPath(roundedRect: element.bounds, byRoundingCorners: corners, cornerRadii: cornerRadii).cgPath
+    XCTAssertEqual(mask?.path, mockPath)
+    XCTAssertTrue(element.contentView.layer.masksToBounds)
+  }
 
 }
-
-// MARK: - Classes with CornerDesignable conformance
-extension AnimatableButton: MockUIViewCornerDesignable {}
-extension AnimatableCheckBox: MockUIViewCornerDesignable {}
-extension AnimatableCollectionViewCell: MockUICollectionViewCellCornerDesignable {}
-extension AnimatableImageView: MockUIViewCornerDesignable {}
-extension AnimatableLabel: MockUIViewCornerDesignable {}
-extension AnimatableScrollView: MockUIViewCornerDesignable {}
-extension AnimatableStackView: MockUIViewCornerDesignable {}
-extension AnimatableTableViewCell: MockUIViewCornerDesignable {}
-extension AnimatableTextField: MockUIViewCornerDesignable {}
-extension AnimatableTextView: MockUIViewCornerDesignable {}
-extension AnimatableView: MockUIViewCornerDesignable {}